+++ apps/src/lab2/views/Lab2Wrapper.tsx
@@ -6,16 +6,24 @@
 // boundary; a fade-in between levels; a loading spinner when a level takes a
 // while to load; and a sad bee when things go wrong.
 
-import React from 'react';
+import React, {useEffect} from 'react';
 import {useSelector} from 'react-redux';
 import classNames from 'classnames';
 import moduleStyles from './Lab2Wrapper.module.scss';
 import ErrorBoundary from '../ErrorBoundary';
-import {LabState, isLabLoading, hasPageError} from '../lab2Redux';
+import {
+  LabState,
+  isLabLoading,
+  hasPageError,
+  setIsShareView,
+} from '../lab2Redux';
 
 import {ErrorFallbackPage, ErrorUI} from './ErrorFallbackPage';
 import Lab2Registry from '../Lab2Registry';
 import Loading from './Loading';
+import {useAppDispatch, useAppSelector} from '@cdo/apps/util/reduxHooks';
+import {getServerLevelId, isShareView} from '../projects/utils';
+import {setCurrentLevelId} from '@cdo/apps/code-studio/progressRedux';
 
 export interface Lab2WrapperProps {
   children: React.ReactNode;
@@ -28,6 +36,28 @@ const Lab2Wrapper: React.FunctionComponent<Lab2WrapperProps> = ({children}) => {
     (state: {lab: LabState}) =>
       state.lab.pageError?.errorMessage || state.lab.pageError?.error?.message
   );
+
+  // Store some server-provided data in redux.
+
+  const dispatch = useAppDispatch();
+  const currentLevelId = useAppSelector(state => state.progress.currentLevelId);
+
+  // Store the level ID provided by App Options in redux if necessary.
+  // This is needed on pages without a header, such as the share view.
+  const serverLevelId = getServerLevelId();
+  useEffect(() => {
+    if (!currentLevelId && serverLevelId) {
+      dispatch(setCurrentLevelId(serverLevelId.toString()));
+    }
+  }, [currentLevelId, serverLevelId, dispatch]);
+
+  // Store whether we are in share view in redux, from App Options.
+  const shareView = isShareView();
+  useEffect(() => {
+    if (shareView !== undefined) {
+      dispatch(setIsShareView(shareView));
+    }
+  }, [shareView, dispatch]);
 
   return (
     <ErrorBoundary
